// enums/enums.ts
export enum SortBy {
  SCORE = "score",
  PRICE = "price",
  VOLUME = "volume",
  RSI = "rsi",
}

export enum SortOrder {
  ASC = "asc",
  DESC = "desc",
}

export enum Indicator {
  RSI_EXTREMES = "rsi_extremes",
  EMA_CROSS = "ema_cross",
  VOLUME_SPIKE = "volume_spike",
  BREAKOUT_VOLUME = "breakout_volume",
  BOLLINGER_SQUEEZE = "bollinger_squeeze",
  ADX_STRONG = "adx_strong",
  PULLBACK_TREAK = "pullback_trend",
}